pull area light tool into its own file
