+++ renderlib/CMakeLists.txt
@@ -16,6 +16,8 @@ target_include_directories(renderlib PUBLIC
 target_sources(renderlib PRIVATE
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/AppScene.h"
+	"${CMAKE_CURRENT_SOURCE_DIR}/AreaLightTool.cpp"
+	"${CMAKE_CURRENT_SOURCE_DIR}/AreaLightTool.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/CCamera.h"
 	"${CMAKE_CURRENT_SOURCE_DIR}/CCamera.cpp"
 	"${CMAKE_CURRENT_SOURCE_DIR}/command.cpp"
