+++ renderlib/Manipulator.h
@@ -104,15 +104,3 @@ private:
   //       some App info line.
   std::string m_identifier;
 };
-
-struct AreaLightTool : ManipulationTool
-{
-
-  AreaLightTool()
-    : ManipulationTool(0)
-  {
-  }
-
-  virtual void action(SceneView& scene, Gesture& gesture) final;
-  virtual void draw(SceneView& scene, Gesture& gesture) final;
-};
